v3.12.2.0
Changelog: https://haproxy-wi.org/changelog.py#3_12_2
