+++ app/overview.py
@@ -36,7 +36,21 @@ try:
 	servers = sql.get_dick_permit()
 except:
 	role = ''
-	pass
+	user = ''
+	users = ''
+	groups = ''
+	roles = ''
+	metrics_master = ''
+	metrics_worker = ''
+	checker_master = ''
+	checker_worker = ''
+	keep_alive = ''
+	api = ''
+	date = ''
+	error = ''
+	versions = ''
+	haproxy_wi_log = ''
+	servers = ''
 
 
 template = template.render(h2 = 1,
