+++ app/templates/add.html
@@ -781,7 +781,7 @@ h3 {
 			In this section you can create and edit black and white lists. And after use them in the HAProxy configs or in the "Add proxy" pages
 		</div>
 		<div id="ajax"></div>
-		<div id="dialog-confirm" title="View certificate " style="display: none;">
+		<div id="dialog-confirm-cert-edit" title="View certificate " style="display: none;">
 			<span><b>Note:</b> Each new address must be specified from a new line</span>
 			<textarea id="edit_lists" style="width: 100%" rows=20></textarea>
 		</div>
