+++ app/templates/base.html
@@ -1,6 +1,6 @@
 <html lang="en">
 	<head>
-		<title>{{title}} - HAproxy-WI</title>
+		<title>{{title}} - HAProxy-WI</title>
 		<meta charset="UTF-8">	
 		<meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate" />
 		<meta http-equiv="Pragma" content="no-cache" />
@@ -105,7 +105,7 @@
 									<li><a href="/app/users.py#groups" title="Actions with groups" class="group head-submenu">Groups</a></li>
 									<li><a href="/app/users.py#servers" title="Actions with servers" class="runtime head-submenu">Servers</a></li>
 									<li><a href="/app/users.py#ssh" title="Manage SSH credentials" class="admin head-submenu">SSH credentials</a></li>
-									<li><a href="/app/users.py#settings" title="HAproxy-WI settings" class="settings head-submenu">Settings</a></li>
+									<li><a href="/app/users.py#settings" title="HAProxy-WI settings" class="settings head-submenu">Settings</a></li>
 									<li><a href="/app/viewlogs.py" title="View internal logs" class="logs head-submenu">Internal logs</a></li>
 									<li><a href="/app/users.py#updatehapwi" title="Update HAProxy-WI" class="upload updatehapwi head-submenu">Update</a></li>
 								</ul>
@@ -260,7 +260,8 @@
 				<a href="https://github.com/Aidaho12/haproxy-wi/" class="footer-link" target="_blank">Github</a>
 				<a href="https://github.com/Aidaho12/haproxy-wi/issues" class="footer-link" target="_blank">Help</a>
 				<a href="https://haproxy-wi.org/contacts.py" class="footer-link" target="_blank">Contact</a>
-				<a href="http://haproxy-wi.org" class="footer-link" target="_blank">About</a>
+				<a href="https://haproxy-wi.org" class="footer-link" target="_blank">About</a>
+				<a href="https://haproxy-wi.org/cloud.py" class="footer-link" target="_blank" title="HAProxy-WI Cloud">Cloud</a>
 			</div>
 		</div>
 	</body>
