+++ app/templates/login.html
@@ -23,7 +23,6 @@ body, .container {
 <center style="margin-top: 15%; background-color: #5d9ceb;">
 	{{error_log}}
 		{{error}}
-		{{db_create}}
 	<div id="login-form" style="padding-top: 40px; padding-bottom: 50px; height: 250px; color: #000;">
 		<span id="logo_span">
 			<img src="/inc/images/logo_login.png" width="330">
