+++ inc/add.js
@@ -2,6 +2,10 @@ var ssl_offloading_var = "http-request set-header X-Forwarded-Port %[dst_port] \
 						"http-request add-header X-Forwarded-Proto https if { ssl_fc } \n"+
 						"redirect scheme https if !{ ssl_fc } \n"
 $( function() {	
+	$('#close').click(function(){
+		$('.alert-success').remove();
+		$('.alert-danger').remove();
+	});
 	$( "#listen-mode-select" ).on('selectmenuchange',function()  {
 		if ($( "#listen-mode-select option:selected" ).val() == "tcp") {
 			$( "#https-listen-span" ).hide("fast");
@@ -1146,4 +1150,84 @@ function change_select_waf(id) {
 			}
 		}
 	} );
+}
+function createList(color) {
+	if(color == 'white') {
+		list = $('#new_whitelist_name').val() 
+	} else {
+		list = $('#new_blacklist_name').val()
+	}
+	$.ajax( {
+		url: "options.py",
+		data: {
+			bwlists_create: list,
+			color: color,
+			group: $('#group').val(),
+			token: $('#token').val()
+		},
+		type: "POST",
+		success: function( data ) {
+			$("#ajax").html(data); 
+			setTimeout(function() {
+						location.reload();
+					}, 2500 );			 
+		}
+	} );	
+}
+function editList(list, color) {
+	$.ajax( {
+		url: "options.py",
+		data: {
+			bwlists: list,
+			color: color,
+			group: $('#group').val(),
+			token: $('#token').val()
+		},
+		type: "POST",
+		success: function( data ) {
+			if (data.indexOf('danger') != '-1') {
+				$("#ajax").html(data);
+			} else {
+				$('.alert-danger').remove();
+				$('#edit_lists').text(data);
+				$( "#dialog-confirm-cert-edit" ).dialog({
+					resizable: false,
+					height: "auto",
+					width: 650,
+					modal: true,
+					title: "Edit "+color+" list "+list,
+					buttons: {
+						"Just save": function() {
+							$( this ).dialog( "close" );	
+							saveList('save', list, color);
+						},
+						"Save and restart": function() {
+							$( this ).dialog( "close" );	
+							saveList('restart', list, color);
+						},
+						Cancel: function() {
+							$( this ).dialog( "close" );
+						}
+					  }
+				});					
+			} 
+		}
+	} );	
+}
+function saveList(action, list, color) {
+	$.ajax( {
+		url: "options.py",
+		data: {
+			bwlists_save: list,
+			bwlists_content: $('#edit_lists').val(),
+			color: color,
+			group: $('#group').val(),
+			bwlists_restart: action,
+			token: $('#token').val()
+		},
+		type: "POST",
+		success: function( data ) {
+			$("#ajax").html(data); 
+		}
+	} );	
 }
