+++ inc/script.js
@@ -601,6 +601,9 @@ $( function() {
 			success: function( data ) {
 				if (data.indexOf('ok') != '-1') {
 					window.location.replace(ref);
+				} else if (data.indexOf('disabled') != '-1') {
+					$('.alert').show();
+					$('.alert').html(data);
 				} else if (data.indexOf('ban') != '-1') {
 					ban();				
 				} 
@@ -795,86 +798,6 @@ function replace_text(id_textarea, text_var) {
 	var end = beg + len_var
 	var text_val = str.substring(0, beg) + str.substring(end, len);
 	$(id_textarea).text(text_val);
-}
-function createList(color) {
-	if(color == 'white') {
-		list = $('#new_whitelist_name').val() 
-	} else {
-		list = $('#new_blacklist_name').val()
-	}
-	$.ajax( {
-		url: "options.py",
-		data: {
-			bwlists_create: list,
-			color: color,
-			group: $('#group').val(),
-			token: $('#token').val()
-		},
-		type: "POST",
-		success: function( data ) {
-			$("#ajax").html(data); 
-			setTimeout(function() {
-						location.reload();
-					}, 2500 );			 
-		}
-	} );	
-}
-function editList(list, color) {
-	$.ajax( {
-		url: "options.py",
-		data: {
-			bwlists: list,
-			color: color,
-			group: $('#group').val(),
-			token: $('#token').val()
-		},
-		type: "POST",
-		success: function( data ) {
-			if (data.indexOf('danger') != '-1') {
-				$("#ajax").html(data);
-			} else {
-				$('.alert-danger').remove();
-				$('#edit_lists').text(data);
-				$( "#dialog-confirm" ).dialog({
-					resizable: false,
-					height: "auto",
-					width: 650,
-					modal: true,
-					title: "Edit "+color+" list "+list,
-					buttons: {
-						"Just save": function() {
-							$( this ).dialog( "close" );	
-							saveList('save', list, color);
-						},
-						"Save and restart": function() {
-							$( this ).dialog( "close" );	
-							saveList('restart', list, color);
-						},
-						Cancel: function() {
-							$( this ).dialog( "close" );
-						}
-					  }
-				});					
-			} 
-		}
-	} );	
-}
-function saveList(action, list, color) {
-	$.ajax( {
-		url: "options.py",
-		data: {
-			bwlists_save: list,
-			bwlists_content: $('#edit_lists').val(),
-			color: color,
-			group: $('#group').val(),
-			bwlists_restart: action,
-			token: $('#token').val()
-		},
-		type: "POST",
-		success: function( data ) {
-			$("#ajax").html(data); 
-		}
-	} );	
 }
 function createHistroy() {
 	try {
